Add edit button and refactored some code
Added an edit button to the details window for making the information on the
window editable, with a text Area showing the date modified.

Moved the code in the set all method of the
ShowAccountDetailsController, wich was suppose to initialize the fields
of the window to the initialize method as this makes more sense.